requirements.txt
fastapi
uvicorn
pypdf
sentence-transformers
faiss-cpu
scikit-learn
numpy